rename file from tool to swithContent
